fix seeder (cancellato, in orario, in ritardo)

+++ database/migrations/2026_04_20_174245_create_trains_table.php
@@ -20,7 +20,7 @@ return new class extends Migration {
             $table->unsignedTinyInteger('totale_carrozze');
             $table->boolean('in_orario');
             $table->boolean('cancellato');
-            $table->unsignedTinyInteger('ritardo')->nullable();
+            $table->unsignedSmallInteger('ritardo')->nullable();
             $table->string('binario')->nullable();
